requirements.txt
django-apscheduler==0.5.1
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0